renamed static to assets
